+++ crates/full-node/sov-sequencer/src/preferred/preferred_blob_sender.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use sov_blob_sender::{BlobInternalId, BlobSender};
+use sov_blob_sender::{BlobInternalId, BlobSender, BlobToSend};
 use sov_blob_storage::{PreferredBatchData, PreferredProofData};
 use sov_db::ledger_db::LedgerDb;
 use sov_rollup_interface::node::da::DaService;
@@ -28,11 +28,7 @@ impl<Da: DaService> PreferredBlobSender<Da> {
         sequence_number: u64,
         blob_id: BlobInternalId,
     ) -> anyhow::Result<()> {
-        let blob = PreferredProofData {
-            sequence_number,
-            data: proof_data.to_vec(),
-        };
-        let blob_bytes = Arc::from(borsh::to_vec(&blob)?);
+        let blob_bytes = proof_bytes(&proof_data, sequence_number)?;
 
         debug!(
             sequence_number,
@@ -48,16 +44,10 @@ impl<Da: DaService> PreferredBlobSender<Da> {
         &mut self,
         batch: PreferredSequencerReadBatch,
     ) -> anyhow::Result<()> {
-        let serialized_batch = borsh::to_vec::<PreferredBatchData>(&PreferredBatchData {
-            sequence_number: batch.sequence_number,
-            visible_slots_to_advance: batch.visible_slots_to_advance,
-            data: batch.txs,
-        })?
-        .into();
+        let blob_id = batch.blob_id;
+        let data = batch_bytes(batch)?;
 
-        self.inner
-            .publish_batch_blob(serialized_batch, batch.blob_id)
-            .await?;
+        self.inner.publish_batch_blob(data, blob_id).await?;
 
         Ok(())
     }
@@ -84,7 +74,54 @@ impl<Da: DaService> PreferredBlobSender<Da> {
                 }
             }
         }
-
         Ok(())
     }
+}
+
+pub fn create_blobs_to_send(
+    completed_blobs: Vec<PreferredSequencerReadBlob>,
+) -> anyhow::Result<Vec<(BlobToSend, BlobInternalId)>> {
+    let mut blobs_to_send = Vec::new();
+
+    for blob in completed_blobs {
+        match blob {
+            PreferredSequencerReadBlob::Batch(batch) => {
+                let blob_id = batch.blob_id;
+                let data = batch_bytes(batch)?;
+                blobs_to_send.push((BlobToSend::Batch { data }, blob_id));
+            }
+            PreferredSequencerReadBlob::Proof {
+                data,
+                sequence_number,
+                blob_id,
+            } => {
+                let data = proof_bytes(&data, sequence_number)?;
+                debug!(
+                    sequence_number,
+                    blob_id, "Dispatching proof blob for publishing"
+                );
+
+                blobs_to_send.push((BlobToSend::Proof { data }, blob_id));
+            }
+        }
+    }
+
+    Ok(blobs_to_send)
+}
+
+fn proof_bytes(proof_data: &[u8], sequence_number: u64) -> anyhow::Result<Arc<[u8]>> {
+    let blob = PreferredProofData {
+        sequence_number,
+        data: proof_data.to_vec(),
+    };
+    Ok(Arc::from(borsh::to_vec(&blob)?))
+}
+
+fn batch_bytes(batch: PreferredSequencerReadBatch) -> anyhow::Result<Arc<[u8]>> {
+    Ok(borsh::to_vec::<PreferredBatchData>(&PreferredBatchData {
+        sequence_number: batch.sequence_number,
+        visible_slots_to_advance: batch.visible_slots_to_advance,
+        data: batch.txs,
+    })?
+    .into())
 }