+++ crates/full-node/sov-sequencer/src/standard/mod.rs
@@ -146,6 +146,7 @@ where
             shutdown_sender,
             Duration::from_secs(config.blob_processing_timeout_secs),
             None,
+            Default::default(),
         )
         .await?;