+++ crates/module-system/sov-modules-api/src/gas/traits.rs
@@ -28,6 +28,7 @@ pub trait GasArray:
     + Clone
     + Send
     + Sync
+    + Unpin
     + PartialEq
     + Eq
     + JsonSchema